Temporarily import package and params modules
This allows gnome-weather to run without the new gjs API.
Unfortunately, as a side effect you need to install the application
to try it, running gnome-weather from the source tree no longer works.

Cherry-picked to master because it is not clear if the new gjs
API will be ever merged upstream.