Util: improve GtkBuilder loader
Accept an optional scope object that is exposed to GtkBuilder as 'scope'.
This allows to use <external-object> to have UI definitions that set
properties of JS derived classes.
Additionally, support connecting signals, using property names on the
scope object.

(XXX: the external-object thing is in the GTK docs, but I saw no code for it...)