package: look for gnome-weather.doap instead of src/
Apparently, it's not uncommon to have src/ in the home directory.
This is still wrong, because I might be running the installed version
from the source tree. The right check is to look at the launched
script and see if it is a descedendant of the current directory
(like ./src/gnome-weather), but that's not possible because gjs
does not include the file as ARGV[0].

Also, fix taking up JS modules when running uninstalled.