ForecastSidebar: allow toggling between today and tomorrow
Remove the heuristics for choosing today or tomorrow in the sidebar,
and always show today if possible. Then allow switching to the next
day and back with the button at the end.
This should solve the ambigous cases close to the end of the day,
when the heuristics would not kick in but little or no useful
information was being shown.