Add infrastructure to build RPMS
Applications should be shipped by upstream, not by distributions.
Ideally, that would be done with some form of bundle, but until
those are possible, RPMs are good too.

This is where the full namespacing shows its strenght: because
everything is "org.gnome.Weather.Application", it's impossible
to have file conflicts with distribution packages (except for
gnome-weather itself), and there is no need for downstream patches.