Build infrastructure: just give up
I say it now, I give up. I tried everything to get a reasonable
environment for working with DBus activated apps, and the requirements
are always contrasting.
So I completely decoupled package.js from service/launcher. The
launcher is simply a shell script that execs the gapplication
tool from glib (much like the desktop file was already doing),
and the actual application is no longer a service. If you want it
to behave like one, you pass --gapplication-service (which is
a special flag interpreted by glib, we can't do local command
line handling in gjs).
This is pretty much the same as gnome-software does, except
the application is still namespaced and installed in libdir,
and the file in $bindir is a shim launcher.
Tests pass and so does coverage. The application can be launched
from the shell (nothing changes in that respect, because it
happens through DBus), and it can be launched from the
command line.

Only regression: gnome-weather --help and gnome-weather --version
no longer do what you would expect. Too bad. --help was useless
anyway, the only option in the list was --version, and we
have an about dialog (plus you should use package version, not
upstream version)