Adding feature to show weather conditions for current location
Use Geoclue to get the current place and find nearest weather station
using it and find_nearest_city method of GWeather-location. Add this
current location if the location does not exist already in the
world_view. Don't add location to 'locations' value of settings. To
handle this use one more column named AUTOMATIC that shows whether
location is added automatically or not. Check this column on removing
the location and if it is true then remove that 'iter' and return.

https://bugzilla.gnome.org/show_bug.cgi?id=704268