Update appdata to the new format
Replace <licence> with <metadata_license> and add <project_license>.

Note that the license of the project is GPLv2+. CC-BY and CC-BY-SA
are listed because they are the license of the backgrounds.
There is some BSD code (util.js), and some MIT code (package.js, params.js),
but that's irrelevant: if you get that code from gnome-weather,
it's all GPLv2+, because it's one application / distribution /
bundle / whatever-thing-you-"convey".