Rename package distribution to org.gnome.Weather
Drop the final Application, which is reserved for the actual
application. This is in preparation for having a
org.gnome.Weather.BackgroundService, in the same package.

This in practice just moves the installed files around (sorry
packagers!). The biggest change is the gettext domain, from
org.gnome.Weather.Application to org.gnome.Weather (because the
background service will also need translations), but it looks
like damned lies will cope just fine (uses the module name).

GSettings schemas, desktop files, appdata files, icons etc have
not been renamed, because they belong to the application proper
(and because renaming them would be a compatibility break).

Finally, the git repository and official tarball is still
gnome-weather. But don't tell anyone ;)