Split between the application and the background service
The main reason for the split is to avoid doing geolocation
in the service that handles the search provider (which causes
the geoclue indicator to appear for every shell search).
In the future, the background service will also handle severe
weather advisories, if enabled (and will be able to do so
across sessions, if launched on the user bus).

And besides, it is clean this way too.