Fix history handling
We need to store the last 5 locations in the order the were
accessed by the user, and move the last location to the front
of the list when it is click. We also need to update the
store on disk when a change happens.

As far as the current location is concerned, the design is
that we store on disk if clicked explicitly, otherwise we
forget it.

https://bugzilla.gnome.org/show_bug.cgi?id=741837