Refactor adding and displaying of locations
Remove the 'show-info' signal and all display related methods
from the shared world model. And have explicit calls to the showInfo
method on the window object used instead.

This fixes a bug where adding a location to the shared model would
show the new location on all open windows. Also make sure to only display
a new current location when we are not displaying another location,
or when we are displaying the previous current location.

https://bugzilla.gnome.org/show_bug.cgi?id=740414