Show tomorrow's forecasts at the end of the day
When opening Weather at the end of the day, it invariably shows the text
"Forecast not available", which is not very friendly to the user who is
looking to see what the forecast is for a few hours ahead. So when
absolutely no forecast data is available for the current day, go ahead
and display tomorrow's forecasts instead, by activating that stack.

Also, leave a warning about what seems to be a bug in the existing code
here.

https://bugzilla.gnome.org/show_bug.cgi?id=777200