searchProvider: Fix a warning
Since gjs 1.54, the native Uint8Array type is used for raw data returned
from introspected functions instead of the custom ByteArray type. For
now gjs overrides Uint8Array's native toString() implementation, which
represents the array as comma-separated digits, with an implementation
that is compatible with the previously used ByteArray, but warns about
the deprecated usage.

Fix that warning by using the recommended replacement.

https://gitlab.gnome.org/GNOME/gnome-weather/merge_requests/18