main: Export location information to gnome-shell
Currently gnome-shell integrates Weather by picking up our settings
and showing a corresponding forecast. However that only works when
gnome-weather is installed on the host, but not from a flatpak, as
the setting is "hidden away" in a container in that case.

To address this, export the relvant information over D-Bus instead,
so that gnome-shell can pick it up without poking at our GSettings
schema.

https://gitlab.gnome.org/GNOME/gnome-shell/issues/1158