Revert "Revert "app: Use find_nearest_city() instead of new_detached()""
This reverts commit 1efc471d9c34be9453b5e457fdebd973f5da632d.

This broke location autodetection. Detached locations have no weather
forecast available.

There's nothing wrong with this commit. Using the nearest city is the
right thing to do when autodetecting location. I reverted this only
because it was committed in tandem with 49e6a52a, which really does need
to stay reverted. But this one is fine.

Fixes #95