WeatherWidget: Use fullcolor icons instead of symbolic ones
Currently, WeatherWidget is using get_symbolic_icon_name () from
GWeather.Info in order to get the name for the large icon.

This is not matching with the design, which is asking for the fullcolor
version of the icon, not the symbolic.

This commit replaces that in order to use get_icon_name (), which
returns the names for the fullcolor icons.

Signed-off-by: Nahuel Gomez Castro <nahual_gomca@outlook.com.ar>