Fix chart's drawing by Y axis
Previously, temperature for the day ranged from 0 to the highest value.
This patch make it range from the lowest to the highest value for that
day. Also, it sets spacing between the chart and entryImage atthe top
and entryTemperatureLabel at the bottom to 18 pixels.

The size and position of entryImage and entryTemperatureLabel are
hardcoded for now, but could be calculated in the future.