weather-widget: reduce scrolled windows minimum content width
Currently, the minimum window width is 384px (this requires hiding the 
minimize/maximize/close buttons, as we do in Phosh).

Reducing the minimum content width allows the window to be resized down 
to 360px, so that it perfectly fits the display of mobile phones such as 
the PinePhone of Librem 5.