app: city: use smaller icons on small displays
The current weather icons imposes a constraint on window width, which
can't be lower than 411px with 128x128 icons.

This patch adds smaller (64x64) icons to the app resources and a
HdySqueezer to select the appropriate icon depending on the requested
window size. Unused GtkImage properties are also removed in the process.

This allows resizing the window down to 360px wide.