window: increase default window width to 760px
After the adoption of libhandy, we should make sure that the default
window size ensures that the window is opened in the preferred mode for
desktops. This means, that the ViewSwitcher should be at the top and
expanded with icons next to the text.

Furthermore, as a side effect, this makes the daily forecast accessible
without scrolling.