window: Make the seach view headerbar flat
This separates the seach and city views a bit more so they don't share
the same headerbar with two completely different contents, making the
code of the main headerbar simpler at the cost of having an extra
headerbar. More importantly, this makes the search view's headerbar flat
and gives it no title, so it completely blends with this views pretty
blue background.